Added events and tmpChange method
Added a tmpChange method to the Resource class. Changes the rate of a resource for a given amount of time. Removed the Currency Object, changed all Currency objects to Resources.

Also added a random event function called every sol. Currently there is a 1 in 15 chance of a dust storm that sets energy rate to 0 for 10 seconds.